% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEpiData.R
\name{GetEpiData}
\alias{GetEpiData}
\title{GetEpiData function
This function read the running epi data online}
\usage{
GetEpiData()
}
\description{
GetEpiData function
This function read the running epi data online
}
